Actually mutate self in scale_differentials with a ref mut binding
This was very hard to track down and was causing differentials to not actually be
scaled since the "diff" binding was just a copy of self.diff that was modified then
dropped (I'm kind of surprised this wasn't linted). This in turn caused the wrong
mipmap level to be used which led to blurry image textures. Also change the perspective
znear to be the same as pbrt although this shouldn't affect anything at all.